+++ java-client/src/main/java/org/opensearch/client/opensearch/_types/mapping/FieldType.java
@@ -111,6 +111,8 @@ public enum FieldType implements JsonEnum {
 
     Shape("shape"),
 
+    XyShape("xy_shape"),
+
     Histogram("histogram"),
 
     ConstantKeyword("constant_keyword"),