+++ java-client/src/test/java/org/opensearch/client/opensearch/_types/query_dsl/XyShapeFieldQueryTest.java
@@ -1,0 +1,14 @@
+package org.opensearch.client.opensearch._types.query_dsl;
+
+import org.junit.Test;
+import org.opensearch.client.opensearch.model.ModelTestCase;
+
+public class XyShapeFieldQueryTest extends ModelTestCase {
+    @Test
+    public void toBuilder() {
+        XyShapeFieldQuery origin = new XyShapeFieldQuery.Builder().build();
+        XyShapeFieldQuery copied = origin.toBuilder().build();
+
+        assertEquals(toJson(copied), toJson(origin));
+    }
+}